+++ inc/Skin_Data.php
@@ -69,4 +69,3 @@ class Skin_Data {
 		}
 	}
 }
-
